requirements.txt
django==1.11.2
Fabric==1.8.0
django-extensions==1.9.6
drf-extensions==0.3.1
jcconv==0.2.3
gunicorn==19.4.3
redis==2.10.5
hiredis==0.2.0
pytz==2013d
django-allauth==0.29.0
django-nose==1.4.4
django-lazysignup==1.0.2  # TODO Remove?
django-model-utils==2.4
rq==0.7.1
django-rq==0.9.2
rq-dashboard
natto-py==0.7.0
twython==3.3.0
djangorestframework==3.7.1
django-filter==0.14.0
markdown==2.6.5
django-crispy-forms==1.5.2
requests[security]==2.12.4
humanize==0.5.1
python-dateutil==2.5.2
djoser==0.5.1
django-silk==1.0.0
raven==5.32.0
python-memcached==1.58
sparkpost==1.3.0
Jinja2==2.8
django-webpack-loader==0.3.3
mecab-python==0.996
https://github.com/neithere/django-autoslug/archive/223771076e63c1b50f75fb2a576f2ea4c6bad514.zip # no pypi release yet.
django-admin-sortable==2.0.21
Pillow==4.0.0
lxml==3.7.3
beautifulsoup4==4.6.0
html5lib==0.999999999
cssselect==1.0.1
feedgen==0.5.1

# django-devserver + optional dependencies
https://github.com/dcramer/django-devserver/archive/master.zip
#django_devserver
sqlparse
werkzeug
guppy
line_profiler

bpython
psycopg2
unidecode

https://github.com/kennethreitz/dj-static/archive/e73fa3bce00ac30b78218d60ece9efe931058a95.zip
https://github.com/aehlke/django-cachecow/archive/master.zip
https://github.com/aehlke/django-snippetscream/archive/master.zip

# For django-catnap.
WebOb==1.0.3

# Deprecated, but used for migration to v2:
django-timezones==0.2
https://github.com/aehlke/django-catnap/archive/master.zip

# Not sure if used anymore?
rfc3339==5
